task2, verified all run commands with ifs.

+++ 2-do_deploy_web_static.py
@@ -17,18 +17,32 @@ def do_deploy(archive_path):
         return False
     else:
         fname = archive_path.split('/')[-1]
-        put(archive_path, "/tmp/")
+        r = put(archive_path, "/tmp/")
+        if r.failed:
+            return False
         path = '/data/web_static/releases/' + fname[:-4]
         r = run('mkdir -p %s' % path)
-        run('tar -xzf /tmp/%s -C /data/web_static/releases/%s/' %
-            (fname, fname[:-4]))
-        run('rm /tmp/%s' % fname)
-        run('mv /data/web_static/releases/%s/web_static/*\
+        if r.failed:
+            return False
+        r = run('tar -xzf /tmp/%s -C /data/web_static/releases/%s/' %
+                (fname, fname[:-4]))
+        if r.failed:
+            return False
+        r = run('rm /tmp/%s' % fname)
+        if r.failed:
+            return False
+        r = run('mv /data/web_static/releases/%s/web_static/*\
         /data/web_static/releases/%s' % (fname[:-4], fname[:-4]))
-        run('rm -rf /data/web_static/releases/%s/web_static' % fname[:-4])
-        run('rm -rf /data/web_static/current')
+        if r.failed:
+            return False
+        r = run('rm -rf /data/web_static/releases/%s/web_static' % fname[:-4])
+        if r.failed:
+            return False
+        r = run('rm -rf /data/web_static/current')
+        if r.failed:
+            return False
         r = run('ln -s /data/web_static/releases/%s /data/web_static/current'
-            % fname[:-4])
+                % fname[:-4])
         if r.failed:
             return False
         return True
